.claude/worktrees/serene-poitras/backend/requirements.txt
fastapi==0.115.0
psycopg2-binary==2.9.9
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.9
websockets==13.1
aiofiles==24.1.0
python-pptx==1.0.2
pymupdf==1.24.11
pillow==10.4.0
openai>=2.0.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
numpy==1.26.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
slowapi>=0.1.9
pydantic[email]>=2.0.0
python-docx>=1.1.0
